@@ -26,7 +26,7 @@
         }
     },
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.1.3",
         "ext-json": "*",
         "ext-pdo": "*",
       "ext-readline": "*"
@@ -35,7 +35,7 @@
         "eftec/validationone": "For keeping and storing the messages"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7 || ^6.5"
+        "phpunit/phpunit": "*"
     }
 }
 
