@@ -1,11 +1,11 @@
 <?php
 /** @noinspection PhpUnusedParameterInspection
- * @noinspection NullCoalescingOperatorCanBeUsedInspection 
- * @noinspection PhpPureAttributeCanBeAddedInspection 
- * @noinspection PhpArrayShapeAttributeCanBeAddedInspection 
- * @noinspection PhpMissingParamTypeInspection 
- * @noinspection AccessModifierPresentedInspection 
- * @noinspection PhpMissingReturnTypeInspection 
+ * @noinspection NullCoalescingOperatorCanBeUsedInspection
+ * @noinspection PhpPureAttributeCanBeAddedInspection
+ * @noinspection PhpArrayShapeAttributeCanBeAddedInspection
+ * @noinspection PhpMissingParamTypeInspection
+ * @noinspection AccessModifierPresentedInspection
+ * @noinspection PhpMissingReturnTypeInspection
  * @noinspection UnknownInspectionInspection
  * @noinspection PhpIncompatibleReturnTypeInspection
  * @noinspection ReturnTypeCanBeDeclaredInspection
@@ -25,7 +25,7 @@ use mysql\repomodel\TableParentModel;
 use Exception;
 
 /**
- * Generated by PdoOne Version 2.14.3 Date generated Sat, 24 Jul 2021 21:29:26 -0400. 
+ * Generated by PdoOne Version 2.14.3 Date generated Sat, 24 Jul 2021 21:29:26 -0400.
  * DO NOT EDIT THIS CODE. Use instead the Repo Class.
  * @copyright (c) Jorge Castro C. MIT License  https://github.com/EFTEC/PdoOne
  * Class AbstractTableParentRepo
@@ -35,7 +35,7 @@ use Exception;
  */
 abstract class AbstractTableParentRepo extends TestDb
 {
-    const TABLE = 'TableParent';    
+    const TABLE = 'TableParent';
     const IDENTITY = 'idtablaparentPK';
     const PK = [
 	    'idtablaparentPK'
@@ -166,15 +166,15 @@ abstract class AbstractTableParentRepo extends TestDb
         }
         return $r;
     }
-    
+
     /**
      * It converts a row returned from the database.<br>
      * If the column is missing then it sets the field as null.
-     * 
+     *
      * @param array $row [ref]
-     */    
+     */
     public static function convertOutputVal(&$row) {
-        if($row===false || $row===null) { 
+        if($row===false || $row===null) {
             return;
         }
 		$row['idtablaparentPK']=isset($row['idtablaparentPK']) ? (int)$row['idtablaparentPK'] : null;
@@ -187,7 +187,7 @@ abstract class AbstractTableParentRepo extends TestDb
 		!isset($row['fieldUnique']) and $row['fieldUnique']=null; // varchar
 		!isset($row['fieldKey']) and $row['fieldKey']=null; // no conversion
 		$row['extracol']=isset($row['extracol']) ? PdoOne::dateConvert($row['extracol'], 'sql', 'human') : null;
-		!isset($row['extracol2']) and $row['extracol2']=null; // 
+		!isset($row['extracol2']) and $row['extracol2']=null; //
 		isset($row['_idchildFK'])
             and $row['_idchildFK']['idtablachildPK']=&$row['idchildFK']; // linked MANYTOONE
 		isset($row['_TableParentExt'])
@@ -198,9 +198,9 @@ abstract class AbstractTableParentRepo extends TestDb
     /**
      * It converts a row to be inserted or updated into the database.<br>
      * If the column is missing then it is ignored and not converted.
-     * 
+     *
      * @param array $row [ref]
-     */    
+     */
     public static function convertInputVal(&$row) {
 		isset($row['idtablaparentPK']) and $row['idtablaparentPK']=(int)$row['idtablaparentPK'];
 		isset($row['idchildFK']) and $row['idchildFK']=(int)$row['idchildFK'];
@@ -274,7 +274,7 @@ abstract class AbstractTableParentRepo extends TestDb
      * self::where(['col']=>['value'])::toList(); // s= string/double/date, i=integer, b=bool
      * self::where(['col=?']=>['value'])::toList(); // s= string/double/date, i=integer, b=bool
      * </pre>
-     * 
+     *
      * @param array|string   $sql =self::factory()
      * @param null|array|int $param
      *
@@ -343,12 +343,12 @@ abstract class AbstractTableParentRepo extends TestDb
     /**
      * It returns all the relational fields by type. '*' returns all types.<br>
      * It doesn't return normal columns.
-     * 
+     *
      * @param string $type=['*','MANYTOONE','ONETOMANY','ONETOONE','MANYTOMANY'][$i]
      *
      * @return string[]
      * @noinspection SlowArrayOperationsInLoopInspection
-     */        
+     */
     public static function getRelations($type='all') {
         $r= [
 		    'MANYTOONE' => [
@@ -368,8 +368,8 @@ abstract class AbstractTableParentRepo extends TestDb
             }
             return $result;
         }
-        return isset($r[$type]) ? $r[$type] : [];  
-    
+        return $r[$type] ?? [];
+
     }
 
     /**
@@ -377,23 +377,23 @@ abstract class AbstractTableParentRepo extends TestDb
      * @param array|null $filterValue (optional) the values of the filter
      * @return array|bool|null
      * @throws Exception
-     */    
+     */
     public static function toList($filter=PdoOne::NULL,$filterValue=null) {
        if(self::$useModel) {
             return TableParentModel::fromArrayMultiple( self::_toList($filter, $filterValue));
         }
         return self::_toList($filter, $filterValue);
     }
-    
+
     /**
      * It sets the recursivity. By default, if we query or modify a value, it operates with the fields of the entity.
      * With recursivity, we could use the recursivity of the fields, for example, loading a MANYTOONE relation<br>
      * <b>Example:</b><br>
      * <pre>
      * self::setRecursive([]); // (default) no use recursivity.
-     * self::setRecursive('*'); // recursive every MANYTOONE,ONETOONE,MANYTOONE and ONETOONE relations (first level) 
+     * self::setRecursive('*'); // recursive every MANYTOONE,ONETOONE,MANYTOONE and ONETOONE relations (first level)
      * self::setRecursive('MANYTOONE'); // recursive all relations of the type MANYTOONE (first level)
-     * self::setRecursive(['_relation1','_relation2']); // recursive only the relations of the first level 
+     * self::setRecursive(['_relation1','_relation2']); // recursive only the relations of the first level
      * self::setRecursive(['_relation1','_relation1/_subrelation1']); //recursive the relations (first and second level)
      * </pre>
      * If array then it uses the values to set the recursivity.<br>
@@ -408,7 +408,7 @@ abstract class AbstractTableParentRepo extends TestDb
         if(is_string($recursive)) {
             $recursive=TableParentRepo::getRelations($recursive);
         }
-        return parent::_setRecursive($recursive); 
+        return parent::_setRecursive($recursive);
     }
 
     /**
@@ -446,7 +446,7 @@ abstract class AbstractTableParentRepo extends TestDb
     public static function first($pk = PdoOne::NULL) {
         if(self::$useModel) {
             return TableParentModel::fromArray(self::_first($pk));
-        } 
+        }
         return self::_first($pk);
     }
 
@@ -478,12 +478,12 @@ abstract class AbstractTableParentRepo extends TestDb
     public static function insert(&$entity,$transactional=true) {
         return self::_insert($entity,$transactional);
     }
-    
+
     /**
-     * It merge a new entity(row) into the database. If the entity exists then it is updated, otherwise the entity is 
+     * It merge a new entity(row) into the database. If the entity exists then it is updated, otherwise the entity is
      * inserted<br>
      * @param array|object $entity        =self::factory()
-     * @param bool         $transactional If true (default) then the operation is transactional   
+     * @param bool         $transactional If true (default) then the operation is transactional
      *
      * @return array|false=self::factory()
      * @throws Exception
@@ -507,7 +507,7 @@ abstract class AbstractTableParentRepo extends TestDb
      * It deletes an entity by the primary key
      *
      * @param array|object $entity =self::factory()
-     * @param bool         $transactional If true (default) then the operation is transactional   
+     * @param bool         $transactional If true (default) then the operation is transactional
      *
      * @return false|int
      * @throws Exception
@@ -520,7 +520,7 @@ abstract class AbstractTableParentRepo extends TestDb
      * It deletes an entity by the primary key.
      *
      * @param array $pk =self::factory()
-     * @param bool  $transactional If true (default) then the operation is transactional   
+     * @param bool  $transactional If true (default) then the operation is transactional
      *
      * @return int|false
      * @throws Exception
@@ -528,7 +528,7 @@ abstract class AbstractTableParentRepo extends TestDb
     public static function deleteById($pk,$transactional=true) {
         return self::_deleteById($pk,$transactional);
     }
-    
+
     /**
      * Returns an array with the default values (0 for numbers, empty for string, and array|null if recursive)
      *
@@ -543,15 +543,15 @@ abstract class AbstractTableParentRepo extends TestDb
         $row= [
 		'idtablaparentPK'=>0,
 		'_TableParentExt'=>(in_array($recursivePrefix.'_TableParentExt',$recursive,true))
-		                            ? TableParentExtRepo::factory(null,$recursivePrefix.'_TableParentExt') 
+		                            ? TableParentExtRepo::factory(null,$recursivePrefix.'_TableParentExt')
 		                            : null, /* ONETOONE! */
 		'_TableParentxCategory'=>(in_array($recursivePrefix.'_TableParentxCategory',$recursive,true))
-		                            ? [] 
+		                            ? []
 		                            : null, /* MANYTOMANY! */
 		'fieldVarchar'=>'',
 		'idchildFK'=>0,
-		'_idchildFK'=>(in_array($recursivePrefix.'_idchildFK',$recursive,true)) 
-		                            ? TableChildRepo::factory(null,$recursivePrefix.'_idchildFK') 
+		'_idchildFK'=>(in_array($recursivePrefix.'_idchildFK',$recursive,true))
+		                            ? TableChildRepo::factory(null,$recursivePrefix.'_idchildFK')
 		                            : null, /* MANYTOONE!! */
 		'idchild2FK'=>0,
 		'fieldInt'=>0,
@@ -567,14 +567,14 @@ abstract class AbstractTableParentRepo extends TestDb
 
         if ($values !== null) {
             $row = array_merge($row, $values);
-        }        
+        }
         return $row;
     }
-    
+
     /**
      * It returns an empty array with null values and no recursivity.
      * @param array|null $values=self::factoryNull()
-     * 
+     *
      * @return array
      */
     public static function factoryNull($values=null) {
@@ -594,8 +594,8 @@ abstract class AbstractTableParentRepo extends TestDb
 		];
         if ($values !== null) {
             $row = array_merge($row, $values);
-        }    
-        return $row;        
+        }
+        return $row;
     }
 
 }
