@@ -5,21 +5,21 @@ namespace eftec;
 interface IPdoOneCache
 {
     /**
-     * It returns the value of the cache. If not found then it must returns false
+     * It returns the value of the cache. If not found then it must return false
      *
-     * @param string $uid    The unique id. It is generate by sha256 based in the query, parameters, type of query
+     * @param string $uid    The unique id. It is generated by sha256 based in the query, parameters, type of query
      *                       and method.
      * @param string|string[] $family [optional] It is the family or group of the cache. It could be used to invalidate
      *                       the whole group. For example, to invalidate all the cache related with a table.
      *
-     * @return mixed|bool If the cache is not found then it must returns false
+     * @return mixed|bool If the cache is not found then it must return false
      */
     function getCache($uid, $family = '');
 
     /**
      * It stores a cache. This method is used internally by PdoOne.<br>
      * 
-     * @param string $uid    The unique id. It is generate by sha256 based in the query, parameters, type of query
+     * @param string $uid    The unique id. It is generated by sha256 based in the query, parameters, type of query
      *                       and method.
      * @param string|string[] $family [optional] It is the family or group of the cache. It could be used to invalidate
      *                       the whole group. For example, to invalidate all the cache related with a table.
@@ -35,7 +35,7 @@ interface IPdoOneCache
     /**
      * Invalidate a single cache or a list of cache based in a single uid or in a family/group of cache.
      *
-     * @param string|string[] $uid The unique id. It is generate by sha256 based in the query, parameters, type of
+     * @param string|string[] $uid The unique id. It is generated by sha256 based in the query, parameters, type of
      *                             query and method.
      * @param string|string[] $family [optional] It is the family or group of the cache. It could be used to invalidate
      *                       the whole group. For example, to invalidate all the cache related with a table.
